dashboard-api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
gspread>=5.12.0
google-auth>=2.25.0
requests>=2.31.0
